add ecdh feature
