+++ Cargo.toml
@@ -26,7 +26,7 @@ lowmemory = ["secp256k1-sys/lowmemory"]
 global-context = ["std", "rand-std"]
 
 [dependencies]
-secp256k1-sys = { version = "0.4.0", default-features = false, path = "./secp256k1-sys" }
+secp256k1-sys = { version = "0.4.0", default-features = false, features = ["ecdh"], path = "./secp256k1-sys" }
 bitcoin_hashes = { version = "0.9", optional = true }
 rand = { version = "0.6", default-features = false, optional = true }
 serde = { version = "1.0", default-features = false, optional = true }
