+++ secp256k1-sys/Cargo.toml
@@ -26,7 +26,7 @@ libc = "0.2"
 
 [features]
 default = ["std"]
-recovery = []
-lowmemory = []
 std = []
-
+lowmemory = []
+ecdh = []
+recovery = []
